refactor(platform): narrow no-new-promise carve-out to image-load helper
Extract readImageDimensions into its own file and move the file-scoped
ccstate/no-new-promise exemption onto the helper so org-general-tab.tsx
stays under the rule.
