+++ turbo/apps/platform/src/views/zero-page/components/org-manage/org-general-tab.tsx
@@ -52,6 +52,7 @@ import {
   saveError$,
   setSaveError$,
 } from "../../../../signals/zero-page/settings/org-manage-tabs-state.ts";
+import { readImageDimensions } from "./read-image-dimensions.ts";
 
 const sectionCardStyle = {
   border: "0.7px solid hsl(var(--gray-400))",
@@ -66,24 +67,6 @@ function extractErrorMessage(
 ): string {
   const body = result.body as { error?: { message?: string } } | undefined;
   return body?.error?.message ?? fallback;
-}
-
-function readImageDimensions(
-  file: File,
-): Promise<{ width: number; height: number } | null> {
-  const url = URL.createObjectURL(file);
-  return new Promise((resolve) => {
-    const img = new Image();
-    img.addEventListener("load", () => {
-      URL.revokeObjectURL(url);
-      resolve({ width: img.naturalWidth, height: img.naturalHeight });
-    });
-    img.addEventListener("error", () => {
-      URL.revokeObjectURL(url);
-      resolve(null);
-    });
-    img.src = url;
-  });
 }
 
 async function uploadLogo(
